src/main/java/net/kemitix/ldapmanager/ui/MainWindow.java
/*
The MIT License (MIT)

Copyright (c) 2016 Paul Campbell

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all
copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
SOFTWARE.
*/

package net.kemitix.ldapmanager.ui;

import com.googlecode.lanterna.gui2.BasicWindow;
import com.googlecode.lanterna.gui2.Panel;
import com.googlecode.lanterna.gui2.Window;
import net.kemitix.ldapmanager.events.ApplicationExitRequest;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.context.annotation.Bean;
import org.springframework.stereotype.Component;

import javax.annotation.PostConstruct;
import java.util.Arrays;

/**
 * The main application window in the UI.
 *
 * @author Paul Campbell (pcampbell@kemitix.net)
 */
@Component
class MainWindow extends BasicWindow {

    private final Panel mainPanel;

    /**
     * Constructor.
     *
     * @param mainPanel The main panel
     */
    @Autowired
    MainWindow(final Panel mainPanel) {
        this.mainPanel = mainPanel;
    }

    /**
     * Initializer.
     */
    @PostConstruct
    public void init() {
        setHints(Arrays.asList(Window.Hint.FULL_SCREEN, Window.Hint.NO_DECORATIONS));
        setComponent(mainPanel);
    }

    /**
     * Listener to close the main UI window.
     *
     * @return the listener
     */
    @Bean
    public ApplicationExitRequest.Listener mainWindowApplicationExitRequestListener() {
        return e -> close();
    }
}
